The branch predated the fix for dropping an existing singular message on a
decode error, which reworked the same function this change rewrote. Resolve in
favour of main's in-place merge: it keeps a partially merged value on error,
which is what the owned decoder leaves behind, whereas decoding into a clone
and discarding it on error restores the pre-merge value and reintroduces the
divergence that fix removed. The clone also cost a deep copy on every repeated
occurrence of a message field, on the success path.

The oneof bug does not need that path at all. Reaching the merge branch means
this member already holds the oneof, so no sibling is set and the clear has
nothing to do. Only the fresh-decode path could destroy a live member, so the
clear moves after the decode there and the rest stands as main has it.

Both sets of tests are kept and both hold: the four oneof cases fail against
main without the reorder, and the five message-merge cases still pass, so the
resolution preserves the earlier fix rather than reverting it.

+++ buffa/src/types.rs
@@ -732,12 +732,18 @@ pub fn encode_shared_bytes<B: AsSharedBytes, S: EncodeSink>(value: &B, buf: &mut
 /// length coming from the [`SizeCache`](crate::SizeCache)) and for packed
 /// repeated fields (the payload loop follows).
 ///
-/// The arguments are `(field_number, len)` — both `u32`, so transposing them
-/// compiles but emits a structurally-valid-but-wrong header.
+/// The arguments are `(field_number, len)` — both plain integers, so
+/// transposing them can compile (integer literals infer either width) but
+/// emits a structurally-valid-but-wrong header.
+///
+/// `len` is `u64`, like all size arithmetic feeding the encode path;
+/// callers holding a `u32` (e.g. from
+/// [`SizeCache::consume_next`](crate::SizeCache::consume_next)) widen with
+/// `u64::from`.
 #[inline(always)]
-pub fn put_len_delimited_header(field_number: u32, len: u32, buf: &mut impl EncodeSink) {
+pub fn put_len_delimited_header(field_number: u32, len: u64, buf: &mut impl EncodeSink) {
     Tag::new(field_number, WireType::LengthDelimited).encode(buf);
-    encode_varint(len as u64, buf);
+    encode_varint(len, buf);
 }
 
 /// Write a group field's `StartGroup` tag.
